requirements.txt
fastapi
uvicorn                    # ASGI server
pydantic
jinja2                     # Templating engine used by FastAPI
httpx
sqlalchemy
passlib                    # Password hashing
starlette
configparser               # For handling config files
itsdangerous               # For session management (used by Starlette)
asyncio
bcrypt                     # For password hashing
confluent-kafka            # Added based on quixstreams' requirement
quixstreams        # For handling streams
python-multipart
kafka
PyJWT
templates
cassandra-driver